test: delete irrelevant tests
