--- src/test/java/gift/study/RestTemplateTest.java
@@ -1,42 +1,0 @@
-package gift.study;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
-import static org.springframework.http.MediaType.APPLICATION_FORM_URLENCODED_VALUE;
-
-import gift.product.controller.KakaoController;
-import gift.product.service.KakaoProperties;
-import java.net.URI;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.RequestEntity;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.web.client.RestTemplate;
-
-public class RestTemplateTest {
-    private final RestTemplate client = new RestTemplateBuilder().build();
-
-
-    private KakaoController kakaoController;
-    private KakaoProperties properties;
-
-    @Test
-    void test1() {
-        var url = "https://kauth.kakao.com/oauth/token";
-        var headers = new HttpHeaders();
-        headers.add(CONTENT_TYPE, APPLICATION_FORM_URLENCODED_VALUE);
-        var body = new LinkedMultiValueMap<String, String>();
-            body.add("grant_type", "authorization_code");
-        body.add("client_id", properties.clientId());
-        body.add("redirect_url", properties.redirectUrl());
-        body.add("code", "");
-        var request = new RequestEntity<>(body, headers, HttpMethod.POST, URI.create(url));
-        var response = client.exchange(request, String.class);
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        System.out.println(request);
-    }
-
-}
